dataActivity

+++ app/src/main/java/com/example/administrator/WeChats/data/Friends.java
@@ -5,7 +5,7 @@ import org.litepal.crud.LitePalSupport;
 public class Friends extends LitePalSupport{
     private int id;
     private String name;
-    private double number;
+    private int number;
     private boolean sex;
 
     public int getId() {
@@ -14,7 +14,7 @@ public class Friends extends LitePalSupport{
     public String getName() {
         return name;
     }
-    public double getNumber() {
+    public int getNumber() {
         return number;
     }
     public boolean isSex() {
@@ -26,7 +26,7 @@ public class Friends extends LitePalSupport{
     public void setName(String name) {
         this.name = name;
     }
-    public void setNumber(double number) {
+    public void setNumber(int number) {
         this.number = number;
     }
     public void setSex(boolean sex) {
